Update to Index.md on behalf of Mozilla. (#387)
We are long past the point where we can pretend that the most important thing about software freedom is the software. We cannot demand better from the internet if we do not demand better from our leaders, our colleagues and ourselves. Along with the Open Source Diversity Community, Outreachy and the Software Conservancy project, we support this petition.