Remove claim regarding "defense of pedophilia"
I'm writing this PR purely in the interest of making this letter's appendix more objective. I don't support or admonish this letter, and I don't want to step on anyone's toes or get involved in a big political mess. If you don't like my PR, you can close it off, but my hope is that it'll be accepted or at least discussed properly.

This PR deletes the following text (and updates the citation numbers to be accurate):

```markdown
He defended pedophilia, in general, in saying that “there is little evidence to justify the widespread assumption that willing participation in pedophilia hurts children.”[9]
```

If we look at the cited reference, we find the following entry on Stallman's blog:

> 04 January 2013 (Pedophilia)
> There is little evidence to justify the [widespread assumption that willing participation in pedophilia hurts children](https://www.theguardian.com/society/2013/jan/03/paedophilia-bringing-dark-desires-light).
> 
> Granted, children may not dare say no to an older relative, or may not realize they could say no; in that case, even if they do not overtly object, the relationship may still feel imposed to them. That's not willing participation, it's imposed participation, a different issue.

I do not agree that Stallman's post constitutes, as the letter's text says, a _"defense of pedophilia"_. He's linking to, and quoting/paraphrasing an article published in The Guardian, which says:

> A liberal professor of psychology who studied in the late 1970s will see things very differently from someone working in child protection, or with convicted sex offenders. There is, astonishingly, not even a full academic consensus on whether consensual paedophilic relations necessarily cause harm.
> [...]
> Even now there is no academic consensus on that fundamental question – as Goode found. Some academics do not dispute the view of Tom O'Carroll, a former chairman of PIE and tireless paedophilia advocate with a conviction for distributing indecent photographs of children following a sting operation, that society's outrage at paedophilic relationships is essentially emotional, irrational, and not justified by science. "It is the quality of the relationship that matters," O'Carroll insists. "If there's no bullying, no coercion, no abuse of power, if the child enters into the relationship voluntarily … the evidence shows there need be no harm."

The article seems to be discussing the surprising lack of scientific consensus concerning some aspects of pedophilia. Anyone could read this article and link to it saying, "it appears that there is little evidence to justify the assumption that willing participation in pedophilia hurts children". This is the point that the The Guardian piece is making, that there is a lack of scientific consensus on this matter.

I don't think that by merely repeating The Guardian's statement, or the statement that followed it in his post, that Stallman actually engaged in what this open letter describes as a _"defense of pedophilia"_, and think it would be more accurate if this claim was removed.

Again, please don't mob me for proposing this change. I'm not trying to upset anyone or to make any kind of defense or admonishment of Stallman. I just think this letter could stand to remove a misleading claim.